RecodedCharIDHash: Fix runtime errors detected by UndefinedBehaviorSanitizer
Fix this runtime error in recodebeam_test and unicharcompress_test:

    src/ccutil/unicharcompress.h:84:27: runtime error:
      left shift of 267 by 28 places cannot be represented in type 'int'

code has up to kMaxCodeLen (9) values, so the highest possible value for
i is 8, and the shift value can reach 7 * 8 = 56.

That requires an uint64_t data type.
size_t would fit for 64 bit hosts, but be too small for 32 bit hosts.

+++ src/ccutil/unicharcompress.h
@@ -78,10 +78,10 @@ class RecodedCharID {
   }
   // Hash functor for RecodedCharID.
   struct RecodedCharIDHash {
-    size_t operator()(const RecodedCharID& code) const {
-      size_t result = 0;
+    uint64_t operator()(const RecodedCharID& code) const {
+      uint64_t result = 0;
       for (int i = 0; i < code.length_; ++i) {
-        result ^= code(i) << (7 * i);
+        result ^= static_cast<uint64_t>(code(i)) << (7 * i);
       }
       return result;
     }
